Implement base auto-generator sub-section for editor level.
